+++ src/advanced/main.advanced.tsx
@@ -9,6 +9,7 @@ import {
   hasProductStock,
   isQuantityInRange,
 } from "../shared";
+import { useProductList } from "./hooks/useProductList";
 
 ReactDOM.createRoot(document.getElementById("app")!).render(
   <React.StrictMode>
@@ -17,13 +18,7 @@ ReactDOM.createRoot(document.getElementById("app")!).render(
 );
 
 function App() {
-  const [productList, setProductList] = useState([
-    { id: "p1", name: "상품1", value: 10000, quantity: 50 },
-    { id: "p2", name: "상품2", value: 20000, quantity: 30 },
-    { id: "p3", name: "상품3", value: 30000, quantity: 20 },
-    { id: "p4", name: "상품4", value: 15000, quantity: 0 },
-    { id: "p5", name: "상품5", value: 25000, quantity: 10 },
-  ]);
+  const { productList, updateProductQuantity } = useProductList();
 
   const [totalAmount, setTotalAmount] = useState(0);
 
@@ -39,16 +34,6 @@ function App() {
     () => Math.floor(totalAmount / 1000),
     [totalAmount],
   );
-
-  const updateProductQuantity = ({ quantityOffset, targetProduct }) => {
-    setProductList((prev) =>
-      prev.map((product) =>
-        product.id === targetProduct.id
-          ? { ...product, quantity: product.quantity + quantityOffset }
-          : product,
-      ),
-    );
-  };
 
   const handleAddButtonClick = () => {
     const selectedProduct = getSelectedProduct({
